+++ src/ddc/dnn/DNNID.h
@@ -26,17 +26,17 @@ class DNNID
         string agent_name;
         string dnn_model_name;
         double start_simulation_time;
-        int dnn_identifier;
+        double dnn_identifier;
 
     public:
         DNNID(); 
-        DNNID(string dnn_model_name, string agent_name, double start_simulation_time, int dnn_identifier);
+        DNNID(string dnn_model_name, string agent_name, double start_simulation_time, double dnn_identifier);
 
         string toString() const;
 
         string getAgentName() const { return this->agent_name; }
         string getDNNModelName() const { return this->dnn_model_name; }
-        int getDNNIdentifier() const { return this->dnn_identifier; }
+        double getDNNIdentifier() const { return this->dnn_identifier; }
         double getStartSimulationTime() const { return this->start_simulation_time; }
 
         bool operator==(const DNNID& other) const;
